requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.3
Django==3.2.5
django-cors-headers==3.7.0
django-countries==7.2.1
django-crontab==0.7.1
djangorestframework==3.12.4
idna==3.2
Pillow==8.3.1
psycopg2-binary==2.9.1
pyasn1==0.4.8
PyJWT==2.1.0
pysmb==1.2.7
python-dotenv==0.18.0
python-http-client==3.3.2
pytz==2021.1
requests==2.26.0
sqlparse==0.4.1
starkbank-ecdsa==1.1.1
urllib3==1.26.6
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.3
Django==3.2.5
django-cors-headers==3.7.0
django-crontab==0.7.1
djangorestframework==3.12.4
idna==3.2
isodate==0.6.1
msrest==0.7.1
oauthlib==3.2.1
Pillow==8.3.1
psycopg2-binary==2.9.1
pyasn1==0.4.8
PyJWT==2.1.0
pysmb==1.2.7
python-dotenv==0.18.0
python-http-client==3.3.2
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.1
starkbank-ecdsa==1.1.1
typing-extensions==4.3.0
urllib3==1.26.6
